requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
blas=2.112=mkl
blas-devel=3.9.0=12_osx64_mkl
bzip2=1.0.8=h0d85af4_4
ca-certificates=2021.10.8=h033912b_0
ffmpeg=4.3=h0a44026_0
freetype=2.10.4=h4cff582_1
gettext=0.19.8.1=hd1a6beb_1008
gmp=6.2.1=h2e338ed_0
gnutls=3.6.13=h756fd2b_1
jbig=2.1=h0d85af4_2003
jpeg=9d=hbcb3906_0
lame=3.100=h35c211d_1001
lcms2=2.12=h577c468_0
lerc=3.0=he49afe7_0
libblas=3.9.0=12_osx64_mkl
libcblas=3.9.0=12_osx64_mkl
libcxx=12.0.1=habf9029_0
libdeflate=1.8=h0d85af4_0
libffi=3.4.2=h0d85af4_5
libgfortran=5.0.0=9_3_0_h6c81a4c_23
libgfortran5=9.3.0=h6c81a4c_23
libiconv=1.16=haf1e3a3_0
liblapack=3.9.0=12_osx64_mkl
liblapacke=3.9.0=12_osx64_mkl
libpng=1.6.37=h7cec526_2
libtiff=4.3.0=hd146c10_2
libuv=1.42.0=h0d85af4_0
libwebp-base=1.2.1=h0d85af4_0
libzlib=1.2.11=h9173be1_1013
llvm-openmp=12.0.1=hda6cdc1_1
lz4-c=1.9.3=he49afe7_1
mkl=2021.4.0=h89fa619_689
mkl-devel=2021.4.0=h694c41f_690
mkl-include=2021.4.0=hf224eb6_689
ncurses=6.2=h2e338ed_4
nettle=3.6=hedd7734_0
numpy=1.21.4=py38h49b9922_0
olefile=0.46=pyh9f0ad1d_1
openh264=2.1.1=hfd3ada9_0
openjpeg=2.4.0=h6e7aa92_1
openssl=3.0.0=h0d85af4_2
pillow=8.4.0=py38hee640a0_0
pip=21.3.1=pyhd8ed1ab_0
python=3.8.12=h43ca1e7_2_cpython
python_abi=3.8=2_cp38
pytorch=1.10.1=py3.8_0
readline=8.1=h05e3726_0
setuptools=59.6.0=py38h50d1736_0
sqlite=3.37.0=h23a322b_0
tbb=2021.4.0=h940c156_1
tk=8.6.11=h5dbffcc_1
torchvision=0.11.2=py38_cpu
typing_extensions=4.0.1=pyha770c72_0
wheel=0.37.0=pyhd8ed1ab_1
xz=5.2.5=haf1e3a3_1
zlib=1.2.11=h9173be1_1013
zstd=1.5.0=h582d3a0_0
